+++ src/controllers/mod.rs
@@ -6,12 +6,13 @@ use std::sync::Arc;
 use crate::repositories::base_Repository::InMemoryBaseRepository;
 use crate::services::base_service::BaseServiceImpl;
 use crate::config::Config;
+use crate::router::Router;
 
 
 pub fn base_routes(config: Arc<Config>) -> impl warp::Filter<Extract = impl warp::Reply, Error = Infallible> +Clone {
   let repository = InMemoryBaseRepository::new();
   let service = BaseServiceImpl::new(repository);
-  let controller = base_controller::BaseController::new(service, Arc::clone(&config));
+  let router = Router::new(service, Arc::clone(&config));
 
-  controller.routes()
+  router.routes()
 }
